requirements.txt
PyYAML
base58